+++ src/app/models/project.js
@@ -1,0 +1,6 @@
+export const Project = (title, checklist) => {
+  return {
+    title,
+    checklist
+  }
+}
