+++ src/app/models/toDo.js
@@ -1,0 +1,9 @@
+export const ToDo = (title, description, dueDate, priority, completed) => {
+  return {
+    title,
+    description,
+    dueDate,
+    priority,
+    completed
+  }
+};
